refactor/Improved folder structure
