+++ packages/nocodb/tests/unit/factory/column.ts
@@ -1,12 +1,12 @@
 import { UITypes } from 'nocodb-sdk';
 import request from 'supertest';
-import Column from '../../../../../src/lib/models/Column';
-import FormViewColumn from '../../../../../src/lib/models/FormViewColumn';
-import GalleryViewColumn from '../../../../../src/lib/models/GalleryViewColumn';
-import GridViewColumn from '../../../../../src/lib/models/GridViewColumn';
-import Model from '../../../../../src/lib/models/Model';
-import Project from '../../../../../src/lib/models/Project';
-import View from '../../../../../src/lib/models/View';
+import Column from '../../../src/lib/models/Column';
+import FormViewColumn from '../../../src/lib/models/FormViewColumn';
+import GalleryViewColumn from '../../../src/lib/models/GalleryViewColumn';
+import GridViewColumn from '../../../src/lib/models/GridViewColumn';
+import Model from '../../../src/lib/models/Model';
+import Project from '../../../src/lib/models/Project';
+import View from '../../../src/lib/models/View';
 
 const defaultColumns = [
   {
