+++ packages/nocodb/tests/unit/tsconfig.json
@@ -55,16 +55,18 @@
       "src/types"
     ]
   },
+  "parserOptions": {
+    "sourceType": "module",
+    "tsconfigRootDir": "./",
+    "project": "./tsconfig.json",
+  },
   "include": [
-    "src/**/*.ts",
-//    "src/lib/xgene/migrations/*.js",
-    "src/**/*.json"
+    "./tests/**/**/**.ts",
+    "./tests/**/**.ts"
+    // "**/*.ts",
+    // "**/*.json"
   ],
   "exclude": [
-    "node_modules/**",
-    "node_modules",
-    "../../../xc-lib-private/**",
-    "../../../xc-lib-private"
   ],
   "compileOnSave": false
 }
